requirements.txt
python_telegram_bot==13.2
numpy==1.16.6
boto3==1.17.10